Add files via upload
Duplicate bills can cause financial discrepancies and increase the risk of fraud. This project aims to develop an automated system that detects duplicate invoices using file hashing techniques. By implementing SHA-256 hashing, the system ensures that no bill is uploaded more than once. The system is designed with a user-friendly frontend and a Python Flask-based backend to handle file uploads and comparison efficiently. The solution is scalable and can be integrated with databases for long-term storage.